Initial reconstruction
Bindings are now independent of the injector by using Dependency and DependencyResolution. 'bindTo' method is gone as it was an unpleasant hack..
