+++ lib/src/module.dart
@@ -1,0 +1,26 @@
+// Copyright (c) 2013, the Dart project authors.  Please see the AUTHORS file
+// for details. All rights reserved. Use of this source code is governed by a
+// BSD-style license that can be found in the LICENSE file.
+
+library dado.module;
+
+/**
+ * A Module is a declaration of bindings that instruct an [Injector] how to
+ * create objects.
+ *
+ * Bindings are declared with members on a Module. The return type of the member
+ * defines what type the binding is for. The kind of member (variable, getter,
+ * method) defines the type of binding:
+ *
+ * * Variables define instance bindings. The type of the variable is bound to
+ *   its value.
+ * * Abstract getters define singleton bindings.
+ * * Abstract methods define unscoped bindings. A new instance is created every
+ *   time [Injector.getInstance] is called.
+ * * A non-abstract method must return instances of its return type. Often
+ *   this will be done by calling [bindTo] with a type that is bound to, and
+ *   then either [Binder.singleton] or [Binder.newInstance] dependeing on
+ *   whether the method is a getter or not. Getters define singletons and should
+ *   call [Binder.singleton], methods should call [Binder.newInstance].
+ */
+class Module {}
